+++ src/api/extrasApi.js
@@ -2,12 +2,16 @@ import axiosClient from "./axiosClient"
 
 const extrasApi = {
   createExtra: (params) => axiosClient.post("api/v1/extra", params),
+
   updateExtra: (extraId, params) => axiosClient.patch(`api/v1/extra/${extraId}`, params),
-  deleteExtra: (extraId) => axiosClient.del(`api/v1/extra/${extraId}`),
+
+  deleteExtra: (extraId) => axiosClient.delete(`api/v1/extra/${extraId}`),
 
   createExtraDetails: (extraId, params) => axiosClient.post(`api/v1/extra/${extraId}/detail`, params),
+
   updateExtraDetails: (extraId, params) => axiosClient.patch(`api/v1/extra/${extraId}`, params),
-  deleteExtraDetails: (extraId) => axiosClient.post(`api/v1/extra/${extraId}`)
+
+  deleteExtraDetails: (extraId) => axiosClient.delete(`api/v1/extra/${extraId}`)
 }
 
 export default extrasApi